anonymiser/requirements.txt
certifi==2024.2.2
confluent-kafka==2.3.0
dnspython==2.6.1
psycopg==3.1.18
pymongo==4.6.2
python-dotenv==1.0.1
typing_extensions==4.10.0
